Get match results

Returns the game results for a given league season(s)

get_match_results(
  country,
  gender,
  season_end_year,
  tier = "1st",
  non_dom_league_url = NA
)

Arguments

country

the three character country code

gender

gender of competition, either "M" or "F"

season_end_year

the year(s) the season concludes

tier

the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on

non_dom_league_url

the URL for Cups and Competitions found at https://fbref.com/en/comps/

Value

returns a dataframe with the results of the competition, season and gender

Examples

# \donttest{
df <- get_match_results(country = c("ITA"), gender = "M", season_end_year = 2021)
# for results from English Championship:
df <- get_match_results(country = "ENG", gender = "M", season_end_year = 2021, tier = "2nd")
# for international friendlies:

# }