Get season team stats

Returns different team season statistics results for a given league season and stat type

get_season_team_stats(country, gender, season_end_year, tier, stat_type)

Arguments

country
the three character country code for all countries

gender
gender of competition, either "M", "F" or both

season_end_year
the year the season(s) concludes

tier
the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on

stat_type
the type of team statistics the user requires

The statistic type options (stat_type) include:

"league_table", "league_table_home_away", "standard", "keeper", "keeper_adv", "shooting", "passing", "passing_types", "goal_shot_creation", "defense" , "possession", "playing_time", "misc"

Value

returns a dataframe of a selected team statistic type for a selected league season

Examples

# \donttest{
get_season_team_stats("ITA", "M", 2021, "1st", "defense")
#>    Competition_Name Gender Country Season_End_Year            Squad
#> 1           Serie A      M     ITA            2021         Atalanta
#> 2           Serie A      M     ITA            2021        Benevento
#> 3           Serie A      M     ITA            2021          Bologna
#> 4           Serie A      M     ITA            2021         Cagliari
#> 5           Serie A      M     ITA            2021          Crotone
#> 6           Serie A      M     ITA            2021       Fiorentina
#> 7           Serie A      M     ITA            2021            Genoa
#> 8           Serie A      M     ITA            2021    Hellas Verona
#> 9           Serie A      M     ITA            2021            Inter
#> 10          Serie A      M     ITA            2021         Juventus
#> 11          Serie A      M     ITA            2021            Lazio
#> 12          Serie A      M     ITA            2021            Milan
#> 13          Serie A      M     ITA            2021           Napoli
#> 14          Serie A      M     ITA            2021            Parma
#> 15          Serie A      M     ITA            2021             Roma
#> 16          Serie A      M     ITA            2021        Sampdoria
#> 17          Serie A      M     ITA            2021         Sassuolo
#> 18          Serie A      M     ITA            2021           Spezia
#> 19          Serie A      M     ITA            2021           Torino
#> 20          Serie A      M     ITA            2021          Udinese
#> 21          Serie A      M     ITA            2021      vs Atalanta
#> 22          Serie A      M     ITA            2021     vs Benevento
#> 23          Serie A      M     ITA            2021       vs Bologna
#> 24          Serie A      M     ITA            2021      vs Cagliari
#> 25          Serie A      M     ITA            2021       vs Crotone
#> 26          Serie A      M     ITA            2021    vs Fiorentina
#> 27          Serie A      M     ITA            2021         vs Genoa
#> 28          Serie A      M     ITA            2021 vs Hellas Verona
#> 29          Serie A      M     ITA            2021         vs Inter
#> 30          Serie A      M     ITA            2021      vs Juventus
#> 31          Serie A      M     ITA            2021         vs Lazio
#> 32          Serie A      M     ITA            2021         vs Milan
#> 33          Serie A      M     ITA            2021        vs Napoli
#> 34          Serie A      M     ITA            2021         vs Parma
#> 35          Serie A      M     ITA            2021          vs Roma
#> 36          Serie A      M     ITA            2021     vs Sampdoria
#> 37          Serie A      M     ITA            2021      vs Sassuolo
#> 38          Serie A      M     ITA            2021        vs Spezia
#> 39          Serie A      M     ITA            2021        vs Torino
#> 40          Serie A      M     ITA            2021       vs Udinese
#>    Team_or_Opponent Num_Players Mins_Per_90 Tkl_Tackles TklW_Tackles
#> 1              team          30          38         672          425
#> 2              team          29          38         595          363
#> 3              team          37          38         572          328
#> 4              team          32          38         629          370
#> 5              team          31          38         602          373
#> 6              team          30          38         578          362
#> 7              team          34          38         641          393
#> 8              team          32          38         542          338
#> 9              team          25          38         541          326
#> 10             team          30          38         626          395
#> 11             team          30          38         656          395
#> 12             team          29          38         679          406
#> 13             team          25          38         620          391
#> 14             team          42          38         636          369
#> 15             team          31          38         558          390
#> 16             team          27          38         637          398
#> 17             team          28          38         594          369
#> 18             team          35          38         630          386
#> 19             team          29          38         595          377
#> 20             team          33          38         592          383
#> 21         opponent          30          38         793          514
#> 22         opponent          29          38         504          312
#> 23         opponent          37          38         672          406
#> 24         opponent          32          38         609          371
#> 25         opponent          31          38         665          410
#> 26         opponent          30          38         733          447
#> 27         opponent          34          38         613          416
#> 28         opponent          32          38         589          378
#> 29         opponent          25          38         516          303
#> 30         opponent          30          38         617          358
#> 31         opponent          30          38         604          397
#> 32         opponent          29          38         593          373
#> 33         opponent          25          38         550          346
#> 34         opponent          42          38         562          359
#> 35         opponent          31          38         562          345
#> 36         opponent          27          38         533          316
#> 37         opponent          28          38         685          406
#> 38         opponent          35          38         533          318
#> 39         opponent          29          38         629          375
#> 40         opponent          33          38         633          387
#>    Def 3rd_Tackles Mid 3rd_Tackles Att 3rd_Tackles Tkl_Vs_Dribbles
#> 1              273             293             106             182
#> 2              344             186              65             218
#> 3              305             195              72             171
#> 4              322             239              68             201
#> 5              333             195              74             198
#> 6              297             208              73             218
#> 7              338             224              79             214
#> 8              262             222              58             165
#> 9              249             205              87             178
#> 10             285             251              90             193
#> 11             318             257              81             231
#> 12             345             258              76             257
#> 13             299             243              78             218
#> 14             359             221              56             232
#> 15             225             257              76             163
#> 16             306             262              69             186
#> 17             301             188             105             197
#> 18             289             258              83             215
#> 19             284             224              87             215
#> 20             325             213              54             207
#> 21             481             251              61             264
#> 22             230             206              68             155
#> 23             324             270              78             238
#> 24             290             230              89             188
#> 25             284             274             107             219
#> 26             377             282              74             199
#> 27             288             247              78             231
#> 28             316             227              46             177
#> 29             288             175              53             164
#> 30             301             221              95             200
#> 31             302             214              88             196
#> 32             301             217              75             213
#> 33             294             189              67             209
#> 34             245             231              86             177
#> 35             302             208              52             226
#> 36             240             209              84             179
#> 37             359             221             105             248
#> 38             260             212              61             148
#> 39             287             257              85             245
#> 40             290             258              85             183
#>    Att_Vs_Dribbles Tkl_percent_Vs_Dribbles Past_Vs_Dribbles Press_Pressures
#> 1              549                    33.2              367            5673
#> 2              613                    35.6              395            5031
#> 3              519                    32.9              348            5460
#> 4              573                    35.1              372            5752
#> 5              535                    37.0              337            5249
#> 6              556                    39.2              338            5179
#> 7              587                    36.5              373            5896
#> 8              451                    36.6              286            4976
#> 9              492                    36.2              314            5079
#> 10             557                    34.6              364            5048
#> 11             696                    33.2              465            5648
#> 12             614                    41.9              357            5848
#> 13             625                    34.9              407            5074
#> 14             648                    35.8              416            5566
#> 15             568                    28.7              405            5570
#> 16             598                    31.1              412            6173
#> 17             515                    38.3              318            4943
#> 18             564                    38.1              349            5325
#> 19             577                    37.3              362            5180
#> 20             579                    35.8              372            5073
#> 21             712                    37.1              448            6254
#> 22             406                    38.2              251            4646
#> 23             648                    36.7              410            5242
#> 24             553                    34.0              365            4822
#> 25             599                    36.6              380            5666
#> 26             541                    36.8              342            5062
#> 27             524                    44.1              293            5052
#> 28             524                    33.8              347            5207
#> 29             443                    37.0              279            5671
#> 30             628                    31.8              428            5650
#> 31             545                    36.0              349            5548
#> 32             663                    32.1              450            5209
#> 33             589                    35.5              380            5734
#> 34             558                    31.7              381            5048
#> 35             641                    35.3              415            5298
#> 36             470                    38.1              291            4795
#> 37             682                    36.4              434            6305
#> 38             460                    32.2              312            5554
#> 39             639                    38.3              394            5396
#> 40             591                    31.0              408            5584
#>    Succ_Pressures _percent_Pressures Def 3rd_Pressures Mid 3rd_Pressures
#> 1            1828               32.2              1497              2650
#> 2            1446               28.7              1971              2122
#> 3            1548               28.4              1944              2242
#> 4            1552               27.0              2095              2516
#> 5            1477               28.1              2035              2190
#> 6            1473               28.4              1814              2248
#> 7            1550               26.3              2161              2538
#> 8            1503               30.2              1580              2251
#> 9            1434               28.2              1648              2069
#> 10           1517               30.1              1648              2183
#> 11           1751               31.0              1872              2504
#> 12           1718               29.4              1928              2655
#> 13           1496               29.5              1608              2109
#> 14           1486               26.7              2162              2369
#> 15           1599               28.7              1614              2644
#> 16           1763               28.6              1989              2707
#> 17           1515               30.6              1571              2027
#> 18           1607               30.2              1594              2456
#> 19           1495               28.9              1789              2256
#> 20           1450               28.6              2030              1990
#> 21           1932               30.9              2667              2532
#> 22           1359               29.3              1428              2091
#> 23           1536               29.3              1875              2231
#> 24           1413               29.3              1618              2077
#> 25           1788               31.6              1552              2557
#> 26           1480               29.2              1862              2271
#> 27           1578               31.2              1459              2307
#> 28           1736               33.3              1872              2325
#> 29           1524               26.9              2097              2306
#> 30           1432               25.3              1974              2368
#> 31           1661               29.9              1879              2357
#> 32           1484               28.5              1893              2128
#> 33           1402               24.5              2019              2387
#> 34           1466               29.0              1549              2259
#> 35           1509               28.5              1794              2279
#> 36           1538               32.1              1599              2083
#> 37           1683               26.7              2222              2626
#> 38           1572               28.3              1659              2647
#> 39           1505               27.9              1701              2460
#> 40           1610               28.8              1831              2435
#>    Att 3rd_Pressures Blocks_Blocks Sh_Blocks ShSv_Blocks Pass_Blocks Int
#> 1               1526           539        93           3         446 501
#> 2                938           551       137           1         414 414
#> 3               1274           653       153           2         500 446
#> 4               1141           574       148           1         426 379
#> 5               1024           580       145           1         435 357
#> 6               1117           510       139           1         371 368
#> 7               1197           592       137           1         455 475
#> 8               1145           594       169           1         425 463
#> 9               1362           529       122           1         407 365
#> 10              1217           543       129           1         414 370
#> 11              1272           563       102           1         461 423
#> 12              1265           645       153           4         492 383
#> 13              1357           527       115           1         412 407
#> 14              1035           596       127           1         469 430
#> 15              1312           499        98           3         401 457
#> 16              1477           662       170           5         492 409
#> 17              1345           575       158           2         417 372
#> 18              1275           530       106           2         424 436
#> 19              1135           554       134           5         420 493
#> 20              1053           619       155           1         464 399
#> 21              1055           651       160           4         491 522
#> 22              1127           522       134           0         388 327
#> 23              1136           605       138           1         467 428
#> 24              1127           555       132           2         423 398
#> 25              1557           501       101           1         400 388
#> 26               929           556        95           2         461 331
#> 27              1286           547       104           5         443 353
#> 28              1010           638       128           3         510 444
#> 29              1268           572       135           0         437 480
#> 30              1308           584       174           1         410 458
#> 31              1312           554       136           3         418 435
#> 32              1188           561       154           4         407 425
#> 33              1328           611       184           2         427 477
#> 34              1240           536       111           2         425 364
#> 35              1225           604       155           1         449 460
#> 36              1113           625       141           1         484 424
#> 37              1457           603       161           1         442 417
#> 38              1248           532       104           2         428 442
#> 39              1235           531       127           2         404 329
#> 40              1318           547       116           1         431 445
#>    Tkl_plus_Int  Clr Err
#> 1          1173  758   8
#> 2          1009  846  13
#> 3          1018  862  16
#> 4          1008  918   9
#> 5           959  867  16
#> 6           946  773   6
#> 7          1116  981  12
#> 8          1005 1015  16
#> 9           906  632   8
#> 10          996  657   7
#> 11         1079  755   7
#> 12         1062  816  11
#> 13         1027  633   9
#> 14         1066  863  10
#> 15         1015  647  12
#> 16         1046  859  10
#> 17          966  762   8
#> 18         1066  744  16
#> 19         1088  882  11
#> 20          991  782  15
#> 21         1315  857  11
#> 22          831  901  12
#> 23         1100  787  11
#> 24         1007  960   7
#> 25         1053  664  12
#> 26         1064  887   8
#> 27          966  759  13
#> 28         1033  892   5
#> 29          996  731  17
#> 30         1075  732  10
#> 31         1039  874  16
#> 32         1018  821   8
#> 33         1027  689  16
#> 34          926  715   6
#> 35         1022  762  12
#> 36          957  949   8
#> 37         1102  701  10
#> 38          975  776  10
#> 39          958  852  12
#> 40         1078  743  16
get_season_team_stats("ITA", "M", 2021, "2nd", "standard")
#>    Competition_Name Gender Country Season_End_Year             Squad
#> 1           Serie B      M     ITA            2021            Ascoli
#> 2           Serie B      M     ITA            2021           Brescia
#> 3           Serie B      M     ITA            2021            Chievo
#> 4           Serie B      M     ITA            2021        Cittadella
#> 5           Serie B      M     ITA            2021           Cosenza
#> 6           Serie B      M     ITA            2021         Cremonese
#> 7           Serie B      M     ITA            2021            Empoli
#> 8           Serie B      M     ITA            2021         Frosinone
#> 9           Serie B      M     ITA            2021             Lecce
#> 10          Serie B      M     ITA            2021             Monza
#> 11          Serie B      M     ITA            2021           Pescara
#> 12          Serie B      M     ITA            2021              Pisa
#> 13          Serie B      M     ITA            2021         Pordenone
#> 14          Serie B      M     ITA            2021          Reggiana
#> 15          Serie B      M     ITA            2021       Salernitana
#> 16          Serie B      M     ITA            2021              SPAL
#> 17          Serie B      M     ITA            2021        US Reggina
#> 18          Serie B      M     ITA            2021           Venezia
#> 19          Serie B      M     ITA            2021    Vicenza Virtus
#> 20          Serie B      M     ITA            2021    Virtus Entella
#> 21          Serie B      M     ITA            2021         vs Ascoli
#> 22          Serie B      M     ITA            2021        vs Brescia
#> 23          Serie B      M     ITA            2021         vs Chievo
#> 24          Serie B      M     ITA            2021     vs Cittadella
#> 25          Serie B      M     ITA            2021        vs Cosenza
#> 26          Serie B      M     ITA            2021      vs Cremonese
#> 27          Serie B      M     ITA            2021         vs Empoli
#> 28          Serie B      M     ITA            2021      vs Frosinone
#> 29          Serie B      M     ITA            2021          vs Lecce
#> 30          Serie B      M     ITA            2021          vs Monza
#> 31          Serie B      M     ITA            2021        vs Pescara
#> 32          Serie B      M     ITA            2021           vs Pisa
#> 33          Serie B      M     ITA            2021      vs Pordenone
#> 34          Serie B      M     ITA            2021       vs Reggiana
#> 35          Serie B      M     ITA            2021    vs Salernitana
#> 36          Serie B      M     ITA            2021           vs SPAL
#> 37          Serie B      M     ITA            2021     vs US Reggina
#> 38          Serie B      M     ITA            2021        vs Venezia
#> 39          Serie B      M     ITA            2021 vs Vicenza Virtus
#> 40          Serie B      M     ITA            2021 vs Virtus Entella
#>    Team_or_Opponent Num_Players  Age Poss MP_Playing_Time Starts_Playing_Time
#> 1              team          42 26.8 50.2              38                 418
#> 2              team          30 25.6 51.3              38                 418
#> 3              team          30 26.4 52.4              38                 418
#> 4              team          27 26.7 54.9              38                 418
#> 5              team          31 27.7 46.9              38                 418
#> 6              team          30 25.9 49.9              38                 418
#> 7              team          27 25.1 54.3              38                 418
#> 8              team          30 27.4 52.0              38                 418
#> 9              team          30 27.8 52.5              38                 418
#> 10             team          33 26.9 56.8              38                 418
#> 11             team          40 26.8 47.6              38                 418
#> 12             team          29 26.8 49.5              38                 418
#> 13             team          29 25.5 43.7              38                 418
#> 14             team          30 26.4 50.1              38                 407
#> 15             team          34 27.1 38.3              38                 407
#> 16             team          31 26.2 52.4              38                 418
#> 17             team          37 27.7 51.4              38                 418
#> 18             team          30 26.1 51.3              38                 418
#> 19             team          32 28.1 47.7              38                 418
#> 20             team          39 26.2 47.6              38                 418
#> 21         opponent          42 26.8 49.8              38                 418
#> 22         opponent          30 26.6 48.7              38                 418
#> 23         opponent          30 26.8 47.6              38                 418
#> 24         opponent          27 26.7 45.1              38                 418
#> 25         opponent          31 26.5 53.1              38                 418
#> 26         opponent          30 26.7 50.1              38                 418
#> 27         opponent          27 26.7 45.7              38                 418
#> 28         opponent          30 27.0 48.0              38                 418
#> 29         opponent          30 26.5 47.5              38                 418
#> 30         opponent          33 26.8 43.2              38                 418
#> 31         opponent          40 26.6 52.4              38                 418
#> 32         opponent          29 26.5 50.5              38                 418
#> 33         opponent          29 26.7 56.3              38                 418
#> 34         opponent          30 26.6 49.9              38                 407
#> 35         opponent          34 26.5 61.7              38                 407
#> 36         opponent          31 26.8 47.6              38                 418
#> 37         opponent          37 26.7 48.6              38                 418
#> 38         opponent          30 26.7 48.7              38                 418
#> 39         opponent          32 26.6 52.3              38                 418
#> 40         opponent          39 26.8 52.4              38                 418
#>    Min_Playing_Time Mins_Per_90_Playing_Time Gls Ast G_minus_PK PK PKatt CrdY
#> 1              3420                       38  37  25         35  2     4  109
#> 2              3420                       38  61  36         56  5     5   92
#> 3              3420                       38  50  36         46  4     5   66
#> 4              3420                       38  47  33         45  2     3  108
#> 5              3420                       38  28  17         25  3     3   98
#> 6              3420                       38  44  26         40  4     4   93
#> 7              3420                       38  66  50         62  4     6   65
#> 8              3420                       38  38  28         34  4     5   86
#> 9              3420                       38  67  47         59  8    11   84
#> 10             3420                       38  51  38         46  5     8   79
#> 11             3420                       38  29  15         27  2     6   91
#> 12             3420                       38  54  33         48  6     8   93
#> 13             3420                       38  39  31         35  4     5   90
#> 14             3330                       37  31  22         29  2     2   95
#> 15             3330                       37  43  26         38  5     8   92
#> 16             3420                       38  44  32         40  4     5   94
#> 17             3420                       38  41  32         39  2     5  106
#> 18             3420                       38  50  31         45  5     5   89
#> 19             3420                       38  48  39         47  1     3  110
#> 20             3420                       38  28  18         23  5     7  104
#> 21             3420                       38  48  37         44  4     5  105
#> 22             3420                       38  52  35         48  4     7   81
#> 23             3420                       38  37  30         35  2     3   99
#> 24             3420                       38  35  23         32  3     4   76
#> 25             3420                       38  46  33         40  6    12   91
#> 26             3420                       38  41  30         40  1     2   75
#> 27             3420                       38  35  23         30  5     6   76
#> 28             3420                       38  41  29         40  1     2   94
#> 29             3420                       38  46  34         45  1     3   97
#> 30             3420                       38  31  18         28  3     3   98
#> 31             3420                       38  58  43         54  4     6   93
#> 32             3420                       38  59  35         52  7     8   98
#> 33             3420                       38  38  25         35  3     3  114
#> 34             3330                       37  54  34         50  4     4   83
#> 35             3330                       37  34  22         32  2     2   90
#> 36             3420                       38  42  31         38  4     7  105
#> 37             3420                       38  45  30         41  4     9  105
#> 38             3420                       38  39  26         35  4     5  101
#> 39             3420                       38  52  37         44  8    10   83
#> 40             3420                       38  63  40         56  7     7   80
#>    CrdR Gls_Per_Minutes Ast_Per_Minutes G_plus_A_Per_Minutes
#> 1     9            0.97            0.66                 1.63
#> 2     6            1.61            0.95                 2.55
#> 3     5            1.32            0.95                 2.26
#> 4     3            1.24            0.87                 2.11
#> 5     9            0.74            0.45                 1.18
#> 6     5            1.16            0.68                 1.84
#> 7     5            1.74            1.32                 3.05
#> 8     1            1.00            0.74                 1.74
#> 9     1            1.76            1.24                 3.00
#> 10    6            1.34            1.00                 2.34
#> 11    6            0.76            0.39                 1.16
#> 12    5            1.42            0.87                 2.29
#> 13    3            1.03            0.82                 1.84
#> 14    7            0.84            0.59                 1.43
#> 15    3            1.16            0.70                 1.86
#> 16    4            1.16            0.84                 2.00
#> 17    5            1.08            0.84                 1.92
#> 18    2            1.32            0.82                 2.13
#> 19    5            1.26            1.03                 2.29
#> 20    5            0.74            0.47                 1.21
#> 21    4            1.26            0.97                 2.24
#> 22    4            1.37            0.92                 2.29
#> 23    4            0.97            0.79                 1.76
#> 24    6            0.92            0.61                 1.53
#> 25    3            1.21            0.87                 2.08
#> 26    3            1.08            0.79                 1.87
#> 27    3            0.92            0.61                 1.53
#> 28    5            1.08            0.76                 1.84
#> 29    6            1.21            0.89                 2.11
#> 30    6            0.82            0.47                 1.29
#> 31    4            1.53            1.13                 2.66
#> 32    3            1.55            0.92                 2.47
#> 33   12            1.00            0.66                 1.66
#> 34    5            1.46            0.92                 2.38
#> 35    3            0.92            0.59                 1.51
#> 36    3            1.11            0.82                 1.92
#> 37    8            1.18            0.79                 1.97
#> 38    6            1.03            0.68                 1.71
#> 39    2            1.37            0.97                 2.34
#> 40    5            1.66            1.05                 2.71
#>    G_minus_PK_Per_Minutes G_plus_A_minus_PK_Per_Minutes
#> 1                    0.92                          1.58
#> 2                    1.47                          2.42
#> 3                    1.21                          2.16
#> 4                    1.18                          2.05
#> 5                    0.66                          1.11
#> 6                    1.05                          1.74
#> 7                    1.63                          2.95
#> 8                    0.89                          1.63
#> 9                    1.55                          2.79
#> 10                   1.21                          2.21
#> 11                   0.71                          1.11
#> 12                   1.26                          2.13
#> 13                   0.92                          1.74
#> 14                   0.78                          1.38
#> 15                   1.03                          1.73
#> 16                   1.05                          1.89
#> 17                   1.03                          1.87
#> 18                   1.18                          2.00
#> 19                   1.24                          2.26
#> 20                   0.61                          1.08
#> 21                   1.16                          2.13
#> 22                   1.26                          2.18
#> 23                   0.92                          1.71
#> 24                   0.84                          1.45
#> 25                   1.05                          1.92
#> 26                   1.05                          1.84
#> 27                   0.79                          1.39
#> 28                   1.05                          1.82
#> 29                   1.18                          2.08
#> 30                   0.74                          1.21
#> 31                   1.42                          2.55
#> 32                   1.37                          2.29
#> 33                   0.92                          1.58
#> 34                   1.35                          2.27
#> 35                   0.86                          1.46
#> 36                   1.00                          1.82
#> 37                   1.08                          1.87
#> 38                   0.92                          1.61
#> 39                   1.16                          2.13
#> 40                   1.47                          2.53
# }