Load match results

Loading version of get_match_results Returns the game results for a given league season(s) from FBref

Usage

load_match_results(country, gender, season_end_year, tier)

Arguments

country

the three character country code

gender

gender of competition, either "M" or "F"

season_end_year

the year(s) the season concludes

tier

the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on

Value

returns a dataframe with the results of the competition, season and gender

Examples

# \donttest{
try({
df <- load_match_results(
country = c("ITA"), gender = "M", season_end_year = 2021, tier = "1st"
)
# for results from English 1st div for men and women:
df <- load_match_results(
country = "ENG", gender = c("M", "F"), season_end_year = 2021, tier = "1st"
)
})
#> → Data last updated 2024-07-18 17:32:43.1420340538025 UTC
#> → Data last updated 2024-07-18 17:31:00.84354496002197 UTC
# }