Team transfer balances

Returns all team's transfer aggregated performances for a chosen league season

tm_team_transfer_balances(country_name, start_year, league_url = NA)

Arguments

country_name
the country of the league's players

start_year
the start year of the season (2020 for the 20/21 season)

league_url
league url from transfermarkt.com. To be used when country_name not avalilable in main function

Value

returns a dataframe of the summarised financial transfer performance of all teams for a league season

Examples

# \donttest{
tm_team_transfer_balances(country_name="England", start_year="2020")
#>    country         league season                   squad expenditure_euros
#> 1  England Premier League   <NA>              Arsenal FC          86000000
#> 2  England Premier League   <NA>             Aston Villa         101350000
#> 3  England Premier League   <NA>  Brighton & Hove Albion          28500000
#> 4  England Premier League   <NA>              Burnley FC           1100000
#> 5  England Premier League   <NA>              Chelsea FC         247200000
#> 6  England Premier League   <NA>          Crystal Palace          22400000
#> 7  England Premier League   <NA>              Everton FC          74870000
#> 8  England Premier League   <NA>               Fulham FC          37250000
#> 9  England Premier League   <NA>            Leeds United         105800000
#> 10 England Premier League   <NA>          Leicester City          58500000
#> 11 England Premier League   <NA>            Liverpool FC          82650000
#> 12 England Premier League   <NA>         Manchester City         161800000
#> 13 England Premier League   <NA>       Manchester United          83800000
#> 14 England Premier League   <NA>        Newcastle United          39000000
#> 15 England Premier League   <NA>        Sheffield United          62700000
#> 16 England Premier League   <NA>          Southampton FC          37300000
#> 17 England Premier League   <NA>       Tottenham Hotspur         110500000
#> 18 England Premier League   <NA>    West Bromwich Albion          42140000
#> 19 England Premier League   <NA>         West Ham United          54700000
#> 20 England Premier League   <NA> Wolverhampton Wanderers          84800000
#>    income_euros avg_age_out avg_age_in
#> 1      19150000        25.8       25.1
#> 2       2770000        27.6       25.3
#> 3      20600000        26.1       23.8
#> 4       2300000        27.4       24.4
#> 5      57400000        25.6       25.1
#> 6      20000000        24.8       24.3
#> 7       4430000        27.0       25.3
#> 8             0        26.6       25.6
#> 9             0        22.7       24.4
#> 10     53770000        26.4       25.4
#> 11     17200000        25.8       22.6
#> 12     76650000        23.5       21.8
#> 13     19500000        25.0       25.5
#> 14       275000        26.2       24.6
#> 15            0        25.3       23.8
#> 16     26300000        26.0       25.5
#> 17     13300000        24.5       24.5
#> 18      8100000        25.8       26.1
#> 19     45410000        28.1       26.3
#> 20     79200000        23.9       23.3
odd_league <- "https://www.transfermarkt.com/league-one/startseite/wettbewerb/GB3"
tm_team_transfer_balances(start_year="2020", league_url=odd_league)
#>    country     league season               squad expenditure_euros income_euros
#> 1  England League One   <NA>  Accrington Stanley                 0       275000
#> 2  England League One   <NA>       AFC Wimbledon                 0            0
#> 3  England League One   <NA>        Blackpool FC                 0            0
#> 4  England League One   <NA>      Bristol Rovers                 0      1400000
#> 5  England League One   <NA>       Burton Albion                 0            0
#> 6  England League One   <NA>   Charlton Athletic            488000      2200000
#> 7  England League One   <NA>     Crewe Alexandra                 0      1060000
#> 8  England League One   <NA>    Doncaster Rovers                 0      1670000
#> 9  England League One   <NA>      Fleetwood Town                 0       390000
#> 10 England League One   <NA>       Gillingham FC                 0            0
#> 11 England League One   <NA>           Hull City            835000      2100000
#> 12 England League One   <NA>        Ipswich Town                 0            0
#> 13 England League One   <NA>        Lincoln City                 0            0
#> 14 England League One   <NA>  Milton Keynes Dons            222000       720000
#> 15 England League One   <NA>    Northampton Town                 0      1100000
#> 16 England League One   <NA>       Oxford United                 0      2700000
#> 17 England League One   <NA> Peterborough United           2500000      5600000
#> 18 England League One   <NA>     Plymouth Argyle                 0            0
#> 19 England League One   <NA>       Portsmouth FC            275000            0
#> 20 England League One   <NA>        Rochdale AFC                 0            0
#> 21 England League One   <NA>     Shrewsbury Town                 0            0
#> 22 England League One   <NA>      Sunderland AFC            338000            0
#> 23 England League One   <NA>        Swindon Town                 0            0
#> 24 England League One   <NA>      Wigan Athletic                 0      8400000
#>    avg_age_out avg_age_in
#> 1         22.6       22.7
#> 2         22.5       21.0
#> 3         24.8       23.8
#> 4         23.7       21.8
#> 5         23.7       23.3
#> 6         24.7       24.8
#> 7         25.1       23.7
#> 8         22.2       22.0
#> 9         25.0       24.3
#> 10        24.5       23.5
#> 11        24.5       23.7
#> 12        22.7       22.8
#> 13        23.9       22.4
#> 14        24.9       24.0
#> 15        25.1       24.5
#> 16        24.1       23.8
#> 17        23.0       20.2
#> 18        23.2       22.7
#> 19        24.2       23.2
#> 20        24.8       20.3
#> 21        24.7       23.8
#> 22        26.4       25.0
#> 23        23.0       22.9
#> 24        25.6       23.6
# }