Get Understat match shot locations

Returns shooting locations for a selected match from Understat.com

understat_match_shots(match_url)

Arguments

match_url
the URL of the match played

Value

returns a dataframe of shooting locations for a selected team season

Examples

# \donttest{
understat_match_shots(match_url = "https://understat.com/match/14789")
#>        id minute      result     X     Y         xG                 player
#> 1  422440      9 MissedShots 0.869 0.441 0.03135269             Semi Ajayi
#> 2  422441     11 MissedShots 0.965 0.460 0.14474539           Okay Yokuslu
#> 3  422442     14        Goal 0.881 0.356 0.23826571        Hal Robson-Kanu
#> 4  422450     35 BlockedShot 0.883 0.336 0.28253871        Hal Robson-Kanu
#> 5  422456     46 MissedShots 0.957 0.590 0.02608211           Kyle Bartley
#> 6  422454     47 MissedShots 0.712 0.403 0.03956936        Matheus Pereira
#> 7  422458     50   SavedShot 0.767 0.590 0.04381445        Matheus Pereira
#> 8  422460     61 MissedShots 0.942 0.626 0.01327414        Darnell Furlong
#> 9  422462     70 MissedShots 0.905 0.491 0.02667309             Semi Ajayi
#> 10 422465     77   SavedShot 0.905 0.392 0.29570183        Hal Robson-Kanu
#> 11 422437      2 BlockedShot 0.782 0.504 0.03780668       Thiago Alcántara
#> 12 422438      3 MissedShots 0.919 0.509 0.04976943     Nathaniel Phillips
#> 13 422439      5 MissedShots 0.926 0.448 0.03913435     Nathaniel Phillips
#> 14 422443     23 MissedShots 0.888 0.462 0.40629628             Sadio Mané
#> 15 422444     26 BlockedShot 0.756 0.642 0.05527206 Trent Alexander-Arnold
#> 16 422445     27 BlockedShot 0.921 0.496 0.12781920          Rhys Williams
#> 17 422446     27   SavedShot 0.942 0.361 0.06873091          Mohamed Salah
#> 18 422447     31 MissedShots 0.817 0.357 0.04077561 Trent Alexander-Arnold
#> 19 422448     32        Goal 0.813 0.350 0.03779993          Mohamed Salah
#> 20 422449     34  ShotOnPost 0.861 0.421 0.34253857        Roberto Firmino
#> 21 422451     39 BlockedShot 0.862 0.285 0.03881568                Fabinho
#> 22 422452     39   SavedShot 0.866 0.288 0.03959307                Fabinho
#> 23 422453     41 BlockedShot 0.780 0.575 0.03130256             Sadio Mané
#> 24 422455     45 BlockedShot 0.794 0.325 0.02635817 Trent Alexander-Arnold
#> 25 422457     49 BlockedShot 0.886 0.400 0.10506576          Mohamed Salah
#> 26 422459     56   SavedShot 0.902 0.673 0.08952659          Mohamed Salah
#> 27 422461     67 BlockedShot 0.813 0.701 0.02821749        Roberto Firmino
#> 28 422463     73 MissedShots 0.906 0.668 0.07385880        Roberto Firmino
#> 29 422464     76 MissedShots 0.725 0.458 0.01714121                Fabinho
#> 30 422466     78 MissedShots 0.797 0.675 0.02576318       Thiago Alcántara
#> 31 422467     82 MissedShots 0.903 0.403 0.12352330 Trent Alexander-Arnold
#> 32 422468     86   SavedShot 0.751 0.370 0.01921204       Thiago Alcántara
#> 33 422469     89 MissedShots 0.851 0.525 0.09763246    Georginio Wijnaldum
#> 34 422470     90 BlockedShot 0.758 0.702 0.01478345    Georginio Wijnaldum
#> 35 422471     91 BlockedShot 0.811 0.445 0.03412251       Thiago Alcántara
#> 36 422472     94        Goal 0.927 0.532 0.10768176                Alisson
#>    home_away player_id      situation season  shotType match_id
#> 1          h      4490       SetPiece   2020      Head    14789
#> 2          h      6932       SetPiece   2020      Head    14789
#> 3          h      1738       OpenPlay   2020  LeftFoot    14789
#> 4          h      1738       OpenPlay   2020  LeftFoot    14789
#> 5          h       964     FromCorner   2020      Head    14789
#> 6          h      7153 DirectFreekick   2020  LeftFoot    14789
#> 7          h      7153       OpenPlay   2020  LeftFoot    14789
#> 8          h      4391     FromCorner   2020      Head    14789
#> 9          h      4490     FromCorner   2020      Head    14789
#> 10         h      1738       OpenPlay   2020  LeftFoot    14789
#> 11         a       229       OpenPlay   2020 RightFoot    14789
#> 12         a      8228     FromCorner   2020      Head    14789
#> 13         a      8228     FromCorner   2020      Head    14789
#> 14         a       838       OpenPlay   2020 RightFoot    14789
#> 15         a      1791 DirectFreekick   2020 RightFoot    14789
#> 16         a      9086     FromCorner   2020 RightFoot    14789
#> 17         a      1250       OpenPlay   2020  LeftFoot    14789
#> 18         a      1791       OpenPlay   2020  LeftFoot    14789
#> 19         a      1250       OpenPlay   2020  LeftFoot    14789
#> 20         a       482       OpenPlay   2020 RightFoot    14789
#> 21         a      3420       OpenPlay   2020 RightFoot    14789
#> 22         a      3420       OpenPlay   2020 RightFoot    14789
#> 23         a       838       OpenPlay   2020 RightFoot    14789
#> 24         a      1791       OpenPlay   2020  LeftFoot    14789
#> 25         a      1250       OpenPlay   2020  LeftFoot    14789
#> 26         a      1250       OpenPlay   2020  LeftFoot    14789
#> 27         a       482       OpenPlay   2020 RightFoot    14789
#> 28         a       482       OpenPlay   2020  LeftFoot    14789
#> 29         a      3420       OpenPlay   2020 RightFoot    14789
#> 30         a       229       OpenPlay   2020 RightFoot    14789
#> 31         a      1791       OpenPlay   2020  LeftFoot    14789
#> 32         a       229       OpenPlay   2020 RightFoot    14789
#> 33         a       771       OpenPlay   2020 RightFoot    14789
#> 34         a       771       OpenPlay   2020 RightFoot    14789
#> 35         a       229     FromCorner   2020 RightFoot    14789
#> 36         a      1257     FromCorner   2020      Head    14789
#>               home_team away_team home_goals away_goals                date
#> 1  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 2  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 3  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 4  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 5  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 6  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 7  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 8  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 9  West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 10 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 11 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 12 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 13 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 14 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 15 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 16 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 17 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 18 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 19 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 20 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 21 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 22 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 23 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 24 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 25 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 26 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 27 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 28 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 29 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 30 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 31 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 32 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 33 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 34 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 35 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#> 36 West Bromwich Albion Liverpool          1          2 2021-05-16 15:30:00
#>           player_assisted lastAction
#> 1         Matheus Pereira      Cross
#> 2         Darnell Furlong    Chipped
#> 3         Matheus Pereira       Pass
#> 4         Darnell Furlong   HeadPass
#> 5         Matheus Pereira     Aerial
#> 6                    <NA>   Standard
#> 7         Hal Robson-Kanu       Pass
#> 8         Matheus Pereira     Aerial
#> 9         Conor Gallagher     Aerial
#> 10             Semi Ajayi    Chipped
#> 11             Sadio Mané     Tackle
#> 12 Trent Alexander-Arnold     Aerial
#> 13       Andrew Robertson     Aerial
#> 14       Thiago Alcántara       Pass
#> 15                   <NA>   Standard
#> 16     Nathaniel Phillips   HeadPass
#> 17 Trent Alexander-Arnold    Chipped
#> 18       Thiago Alcántara       Pass
#> 19             Sadio Mané       Pass
#> 20                   <NA>       None
#> 21       Thiago Alcántara       Pass
#> 22          Mohamed Salah       Pass
#> 23       Thiago Alcántara       Pass
#> 24          Mohamed Salah       Pass
#> 25 Trent Alexander-Arnold       Pass
#> 26 Trent Alexander-Arnold       Pass
#> 27          Mohamed Salah       Pass
#> 28       Thiago Alcántara       Pass
#> 29        Xherdan Shaqiri       Pass
#> 30        Xherdan Shaqiri       Pass
#> 31                Fabinho     TakeOn
#> 32        Roberto Firmino       Pass
#> 33          Mohamed Salah       Pass
#> 34       Andrew Robertson       Pass
#> 35    Georginio Wijnaldum       Pass
#> 36 Trent Alexander-Arnold      Cross
# }