Get Understat match stats table data

Returns the Stats values for a selected match from Understat.com.

Usage

understat_match_stats(match_url)

Arguments

match_url

A `character` string with the URL of the match played.

Value

returns a `data.frame` with data from the stats table for the match.

Details

For `draw_chances`, `home_chances` and `away_chances`, values below 10