Get season team stats

Returns different team season statistics results for a given league season and stat type

get_season_team_stats(
  country,
  gender,
  season_end_year,
  tier,
  stat_type,
  time_pause = 2
)

Arguments

country

the three character country code for all countries

gender

gender of competition, either "M", "F" or both

season_end_year

the year the season(s) concludes

tier

the tier of the league, ie '1st' for the EPL or '2nd' for the Championship and so on

stat_type

the type of team statistics the user requires

time_pause

the wait time (in seconds) between page loads

The statistic type options (stat_type) include:

"league_table", "league_table_home_away", "standard", "keeper", "keeper_adv", "shooting", "passing", "passing_types", "goal_shot_creation", "defense" , "possession", "playing_time", "misc"

Value

returns a dataframe of a selected team statistic type for a selected league season

Examples

# \donttest{
get_season_team_stats("ITA", "M", 2021, "1st", "defense")
#>    Competition_Name Gender Country Season_End_Year            Squad
#> 1           Serie A      M     ITA            2021         Atalanta
#> 2           Serie A      M     ITA            2021        Benevento
#> 3           Serie A      M     ITA            2021          Bologna
#> 4           Serie A      M     ITA            2021         Cagliari
#> 5           Serie A      M     ITA            2021          Crotone
#> 6           Serie A      M     ITA            2021       Fiorentina
#> 7           Serie A      M     ITA            2021            Genoa
#> 8           Serie A      M     ITA            2021    Hellas Verona
#> 9           Serie A      M     ITA            2021            Inter
#> 10          Serie A      M     ITA            2021         Juventus
#> 11          Serie A      M     ITA            2021            Lazio
#> 12          Serie A      M     ITA            2021            Milan
#> 13          Serie A      M     ITA            2021           Napoli
#> 14          Serie A      M     ITA            2021            Parma
#> 15          Serie A      M     ITA            2021             Roma
#> 16          Serie A      M     ITA            2021        Sampdoria
#> 17          Serie A      M     ITA            2021         Sassuolo
#> 18          Serie A      M     ITA            2021           Spezia
#> 19          Serie A      M     ITA            2021           Torino
#> 20          Serie A      M     ITA            2021          Udinese
#> 21          Serie A      M     ITA            2021      vs Atalanta
#> 22          Serie A      M     ITA            2021     vs Benevento
#> 23          Serie A      M     ITA            2021       vs Bologna
#> 24          Serie A      M     ITA            2021      vs Cagliari
#> 25          Serie A      M     ITA            2021       vs Crotone
#> 26          Serie A      M     ITA            2021    vs Fiorentina
#> 27          Serie A      M     ITA            2021         vs Genoa
#> 28          Serie A      M     ITA            2021 vs Hellas Verona
#> 29          Serie A      M     ITA            2021         vs Inter
#> 30          Serie A      M     ITA            2021      vs Juventus
#> 31          Serie A      M     ITA            2021         vs Lazio
#> 32          Serie A      M     ITA            2021         vs Milan
#> 33          Serie A      M     ITA            2021        vs Napoli
#> 34          Serie A      M     ITA            2021         vs Parma
#> 35          Serie A      M     ITA            2021          vs Roma
#> 36          Serie A      M     ITA            2021     vs Sampdoria
#> 37          Serie A      M     ITA            2021      vs Sassuolo
#> 38          Serie A      M     ITA            2021        vs Spezia
#> 39          Serie A      M     ITA            2021        vs Torino
#> 40          Serie A      M     ITA            2021       vs Udinese
#>    Team_or_Opponent Num_Players Mins_Per_90 Tkl_Tackles TklW_Tackles
#> 1              team          30          38         672          425
#> 2              team          29          38         595          363
#> 3              team          37          38         572          328
#> 4              team          32          38         629          370
#> 5              team          31          38         603          374
#> 6              team          30          38         578          362
#> 7              team          34          38         641          394
#> 8              team          32          38         542          338
#> 9              team          25          38         541          326
#> 10             team          30          38         626          395
#> 11             team          30          38         656          395
#> 12             team          29          38         679          406
#> 13             team          25          38         620          391
#> 14             team          42          38         636          369
#> 15             team          31          38         558          390
#> 16             team          27          38         637          398
#> 17             team          28          38         594          369
#> 18             team          35          38         630          386
#> 19             team          29          38         595          377
#> 20             team          33          38         593          384
#> 21         opponent          30          38         793          514
#> 22         opponent          29          38         504          312
#> 23         opponent          37          38         672          406
#> 24         opponent          32          38         609          372
#> 25         opponent          31          38         666          411
#> 26         opponent          30          38         733          447
#> 27         opponent          34          38         613          416
#> 28         opponent          32          38         589          378
#> 29         opponent          25          38         516          303
#> 30         opponent          30          38         617          358
#> 31         opponent          30          38         604          397
#> 32         opponent          29          38         593          373
#> 33         opponent          25          38         550          346
#> 34         opponent          42          38         562          359
#> 35         opponent          31          38         562          345
#> 36         opponent          27          38         533          316
#> 37         opponent          28          38         685          406
#> 38         opponent          35          38         533          318
#> 39         opponent          29          38         629          375
#> 40         opponent          33          38         634          388
#>    Def 3rd_Tackles Mid 3rd_Tackles Att 3rd_Tackles Tkl_Vs_Dribbles
#> 1              273             293             106             182
#> 2              344             186              65             218
#> 3              305             195              72             171
#> 4              322             239              68             200
#> 5              335             194              74             198
#> 6              297             208              73             218
#> 7              338             224              79             214
#> 8              262             222              58             165
#> 9              249             205              87             178
#> 10             285             251              90             193
#> 11             318             257              81             231
#> 12             345             258              76             257
#> 13             299             243              78             218
#> 14             359             221              56             232
#> 15             225             257              76             163
#> 16             306             262              69             186
#> 17             301             188             105             197
#> 18             289             258              83             215
#> 19             284             224              87             215
#> 20             325             214              54             208
#> 21             481             251              61             264
#> 22             230             206              68             155
#> 23             324             270              78             238
#> 24             290             230              89             188
#> 25             284             275             107             220
#> 26             377             282              74             199
#> 27             288             247              78             230
#> 28             316             227              46             177
#> 29             288             175              53             164
#> 30             301             221              95             200
#> 31             302             214              88             196
#> 32             301             217              75             213
#> 33             294             189              67             209
#> 34             245             231              86             177
#> 35             302             208              52             226
#> 36             240             209              84             179
#> 37             359             221             105             248
#> 38             260             212              61             148
#> 39             287             257              85             245
#> 40             292             257              85             183
#>    Att_Vs_Dribbles Tkl_percent_Vs_Dribbles Past_Vs_Dribbles Press_Pressures
#> 1              549                    33.2              367            5673
#> 2              613                    35.6              395            5031
#> 3              519                    32.9              348            5463
#> 4              571                    35.0              371            5752
#> 5              535                    37.0              337            5250
#> 6              556                    39.2              338            5179
#> 7              587                    36.5              373            5893
#> 8              451                    36.6              286            4976
#> 9              492                    36.2              314            5079
#> 10             557                    34.6              364            5048
#> 11             696                    33.2              465            5644
#> 12             614                    41.9              357            5848
#> 13             625                    34.9              407            5074
#> 14             648                    35.8              416            5566
#> 15             568                    28.7              405            5570
#> 16             597                    31.2              411            6173
#> 17             515                    38.3              318            4943
#> 18             564                    38.1              349            5325
#> 19             577                    37.3              362            5179
#> 20             583                    35.7              375            5068
#> 21             712                    37.1              448            6254
#> 22             406                    38.2              251            4646
#> 23             648                    36.7              410            5242
#> 24             553                    34.0              365            4819
#> 25             603                    36.5              383            5661
#> 26             541                    36.8              342            5065
#> 27             522                    44.1              292            5052
#> 28             524                    33.8              347            5206
#> 29             443                    37.0              279            5671
#> 30             628                    31.8              428            5650
#> 31             545                    36.0              349            5548
#> 32             663                    32.1              450            5209
#> 33             589                    35.5              380            5734
#> 34             557                    31.8              380            5048
#> 35             641                    35.3              415            5299
#> 36             470                    38.1              291            4795
#> 37             682                    36.4              434            6301
#> 38             460                    32.2              312            5554
#> 39             639                    38.3              394            5396
#> 40             591                    31.0              408            5584
#>    Succ_Pressures _percent_Pressures Def 3rd_Pressures Mid 3rd_Pressures
#> 1            1828               32.2              1497              2650
#> 2            1446               28.7              1971              2122
#> 3            1549               28.4              1946              2243
#> 4            1552               27.0              2095              2516
#> 5            1477               28.1              2035              2191
#> 6            1473               28.4              1814              2248
#> 7            1550               26.3              2162              2537
#> 8            1503               30.2              1580              2251
#> 9            1434               28.2              1648              2069
#> 10           1517               30.1              1648              2183
#> 11           1750               31.0              1872              2501
#> 12           1718               29.4              1928              2655
#> 13           1496               29.5              1608              2109
#> 14           1486               26.7              2162              2369
#> 15           1599               28.7              1614              2644
#> 16           1763               28.6              1989              2707
#> 17           1515               30.6              1571              2027
#> 18           1607               30.2              1594              2456
#> 19           1494               28.8              1789              2256
#> 20           1447               28.6              2032              1989
#> 21           1932               30.9              2667              2532
#> 22           1359               29.3              1428              2091
#> 23           1536               29.3              1875              2231
#> 24           1413               29.3              1619              2076
#> 25           1785               31.5              1554              2556
#> 26           1481               29.2              1864              2272
#> 27           1578               31.2              1459              2307
#> 28           1735               33.3              1872              2325
#> 29           1524               26.9              2097              2306
#> 30           1432               25.3              1974              2368
#> 31           1661               29.9              1879              2357
#> 32           1484               28.5              1893              2128
#> 33           1402               24.5              2019              2387
#> 34           1466               29.0              1549              2259
#> 35           1509               28.5              1794              2280
#> 36           1538               32.1              1599              2083
#> 37           1682               26.7              2222              2623
#> 38           1572               28.3              1659              2647
#> 39           1505               27.9              1701              2460
#> 40           1610               28.8              1831              2435
#>    Att 3rd_Pressures Blocks_Blocks Sh_Blocks ShSv_Blocks Pass_Blocks Int
#> 1               1526           539        93           3         446 501
#> 2                938           551       137           1         414 414
#> 3               1274           653       153           2         500 446
#> 4               1141           574       148           1         426 379
#> 5               1024           580       145           1         435 357
#> 6               1117           510       139           1         371 368
#> 7               1194           592       137           1         455 476
#> 8               1145           594       169           1         425 463
#> 9               1362           529       122           1         407 365
#> 10              1217           543       129           1         414 370
#> 11              1271           563       102           1         461 424
#> 12              1265           645       153           4         492 383
#> 13              1357           527       115           1         412 407
#> 14              1035           596       127           1         469 430
#> 15              1312           499        98           3         401 457
#> 16              1477           661       170           5         491 410
#> 17              1345           575       158           2         417 372
#> 18              1275           530       106           2         424 436
#> 19              1134           554       134           5         420 493
#> 20              1047           620       155           1         465 399
#> 21              1055           651       160           4         491 522
#> 22              1127           522       134           0         388 327
#> 23              1136           605       138           1         467 428
#> 24              1124           555       132           2         423 399
#> 25              1551           502       101           1         401 388
#> 26               929           556        95           2         461 331
#> 27              1286           547       104           5         443 353
#> 28              1009           638       128           3         510 444
#> 29              1268           572       135           0         437 480
#> 30              1308           584       174           1         410 458
#> 31              1312           554       136           3         418 435
#> 32              1188           561       154           4         407 425
#> 33              1328           611       184           2         427 477
#> 34              1240           535       111           2         424 365
#> 35              1225           604       155           1         449 460
#> 36              1113           625       141           1         484 424
#> 37              1456           603       161           1         442 418
#> 38              1248           532       104           2         428 442
#> 39              1235           531       127           2         404 329
#> 40              1318           547       116           1         431 445
#>    Tkl_plus_Int  Clr Err
#> 1          1173  758   8
#> 2          1009  846  13
#> 3          1018  862  16
#> 4          1008  918   9
#> 5           960  866  16
#> 6           946  773   6
#> 7          1117  980  12
#> 8          1005 1015  16
#> 9           906  632   8
#> 10          996  657   7
#> 11         1080  755   7
#> 12         1062  816  11
#> 13         1027  633   9
#> 14         1066  863  10
#> 15         1015  647  12
#> 16         1047  859  10
#> 17          966  762   8
#> 18         1066  744  16
#> 19         1088  882  11
#> 20          992  782  15
#> 21         1315  857  11
#> 22          831  901  12
#> 23         1100  787  11
#> 24         1008  959   7
#> 25         1054  664  12
#> 26         1064  887   8
#> 27          966  759  13
#> 28         1033  892   5
#> 29          996  731  17
#> 30         1075  732  10
#> 31         1039  874  16
#> 32         1018  821   8
#> 33         1027  689  16
#> 34          927  715   6
#> 35         1022  762  12
#> 36          957  949   8
#> 37         1103  701  10
#> 38          975  776  10
#> 39          958  852  12
#> 40         1079  742  16
# }